# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Proyecto Carga Horaria - API",
        "description": "Documentación completa de la API del sistema de gestión de carga horaria docente",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/schedules": {
            "get": {
                "tags": [
                    "Horarios"
                ],
                "summary": "CU13 - Listar horarios",
                "operationId": "c2791edf6c0550c67fac652cad98e09e",
                "parameters": [
                    {
                        "name": "group_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "teacher_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de horarios"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Horarios"
                ],
                "summary": "CU13 - Crear horario",
                "operationId": "cec681726db3f091faae4410db8ee14c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "group_id",
                                    "day_of_week",
                                    "start_time",
                                    "end_time"
                                ],
                                "properties": {
                                    "group_id": {
                                        "type": "integer"
                                    },
                                    "room_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "teacher_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "day_of_week": {
                                        "type": "string",
                                        "example": "lunes"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "example": "08:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "example": "10:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Horario creado"
                    },
                    "409": {
                        "description": "Conflicto de horario"
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
        "/api/schedules/weekly": {
            "get": {
                "tags": [
                    "Horarios"
                ],
                "summary": "Ver horarios semanales agrupados por día",
                "operationId": "14493faca0abc670fdaaddd055f338bd",
                "parameters": [
                    {
                        "name": "teacher_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "group_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Horarios agrupados por día"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
        "/api/schedules/export": {
            "get": {
                "tags": [
                    "Horarios"
                ],
                "summary": "CU16 - Exportar horarios a Excel",
                "description": "Exporta los horarios en formato Excel (.xlsx)",
                "operationId": "1025b846ee23de85cfe8613e582a4415",
                "parameters": [
                    {
                        "name": "teacher_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "group_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Archivo Excel",
                        "content": {
                            "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {}
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
        "/api/schedules/export.pdf": {
            "get": {
                "tags": [
                    "Horarios"
                ],
                "summary": "CU16 - Exportar horarios a PDF",
                "description": "Exporta los horarios en formato PDF",
                "operationId": "c052f4b01d2be903be422fe1425c0646",
                "parameters": [
                    {
                        "name": "teacher_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "group_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Archivo PDF",
                        "content": {
                            "application/pdf": {}
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
...
        "/api/schedules/{id}/cancel": {
            "post": {
                "tags": [
                    "Horarios"
                ],
                "summary": "CU19 - Anular clase o cambiar a modalidad virtual",
                "operationId": "d084344f105aef3d3f36dfb72cc24a14",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "mode"
                                ],
                                "properties": {
                                    "mode": {
                                        "type": "string",
                                        "enum": [
                                            "cancelled",
                                            "virtual"
                                        ]
                                    },
                                    "reason": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cancelación registrada"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
        "/api/schedules/{id}/qrcode": {
            "get": {
                "tags": [
                    "Horarios"
                ],
                "summary": "CU18 - Generar código QR firmado para un horario",
                "description": "Genera un código QR con token firmado HMAC-SHA256 para registro de asistencia",
                "operationId": "c9ab73b257d3b0da58187ad95d42ca15",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del horario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Imagen PNG del código QR",
                        "content": {
                            "image/png": {}
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Horario no encontrado"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "name": "laravel_session",
                "in": "cookie"
            }
        }
    },
...
}